packages/faasit-examples/examples/dnn-train/code/requirements.txt

# pytorch
## https://pytorch.org/get-started/locally/
## pip3 install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
## pip3 install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu

pytorch-ignite>=0.4.0

gpustat<1.0; sys_platform != 'darwin'
pynvml>=11.5.0; sys_platform != 'darwin'
psutil>=5.9.6

# used for datasets
scipy>=1.10.1
boto3>=1.17.0  # s3 client
opencv-python-headless>=4.8

# common
pydantic>=2.5.0
fire>=0.5.0
cachetools>=5.3.0
python-snappy>=0.7.1

httpx>=0.20.0
aiohttp>=3.7.4

pyyaml>=6.0
tqdm>=4.62.3